+++ client/CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+* Added step decorator schemas.
+
 ## [0.4.2] - 2021-09-01
 
 ### Changed